♻️ refactor: update SET command with options and enhance store data structure
