+++ internal/core/store/store.go
@@ -11,17 +11,19 @@ import (
 	"github.com/andrelcunha/goodiesdb/internal/utils/slice"
 )
 
+var ErrNoSuchKey = fmt.Errorf("no such key")
+
 type Store struct {
-	data    []map[string]Value
+	data    []map[string]*Value
 	mu      sync.RWMutex
 	aofChan chan string
 }
 
 // NewStore creates a new store
 func NewStore(aofChan chan string) *Store {
-	data := make([]map[string]Value, 16)
+	data := make([]map[string]*Value, 16)
 	for i := range data {
-		data[i] = make(map[string]Value)
+		data[i] = make(map[string]*Value)
 	}
 	return &Store{
 		data:    data,
@@ -35,19 +37,17 @@ func (s *Store) Count() int {
 	return len(s.data)
 }
 
-var ErrNoSuchKey = fmt.Errorf("no such key")
-
 // GetSnapshot returns a snapshot of store data for persistence
 // This is safe to call as it returns a copy
-func (s *Store) GetSnapshot() []map[string]Value {
+func (s *Store) GetSnapshot() []map[string]*Value {
 	s.mu.RLock()
 	defer s.mu.RUnlock()
 
 	// Create deep copies to avoid data races
-	dataCopy := make([]map[string]Value, len(s.data))
+	dataCopy := make([]map[string]*Value, len(s.data))
 
 	for i := range s.data {
-		dataCopy[i] = make(map[string]Value)
+		dataCopy[i] = make(map[string]*Value)
 
 		for k, v := range s.data[i] {
 			dataCopy[i][k] = v
@@ -59,7 +59,7 @@ func (s *Store) GetSnapshot() []map[string]Value {
 }
 
 // RestoreFromSnapshot restores store data from persistence
-func (s *Store) RestoreFromSnapshot(data []map[string]Value) {
+func (s *Store) RestoreFromSnapshot(data []map[string]*Value) {
 	s.mu.Lock()
 	defer s.mu.Unlock()
 
@@ -103,7 +103,7 @@ func (s *Store) GetList(dbIndex int, key string) []any {
 func (s *Store) SetRawValue(dbIndex int, key string, value interface{}) {
 	s.mu.Lock()
 	defer s.mu.Unlock()
-	data := Value{Data: value}
+	data := &Value{Data: value}
 	s.data[dbIndex][key] = data
 }
 
@@ -111,101 +111,15 @@ func (s *Store) AOFChannel() chan string {
 	return s.aofChan
 }
 
-// Set sets the value for a key
-// Consider ret
-func (s *Store) Set(dbIndex int, key string, rawValue any, args ...string) (bool, error) {
-	setOptions, err := parseSetOptions(args)
-	if err != nil {
-		return false, err
-	}
-	// Handle NX and XX options
-	if setOptions.NX && s.Exists(dbIndex, key) > 0 {
-		return false, nil
-	}
-	if setOptions.XX && s.Exists(dbIndex, key) == 0 {
-		return false, nil
-	}
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	// write to AOF before setting the value (WAL)
-	s.aofChan <- fmt.Sprintf("SET %d %s %v", dbIndex, key, rawValue)
-	var value *Value
-	switch v := rawValue.(type) {
-	case string:
-		value = NewStringValue(v)
-	case []any:
-		value = NewListValue(v)
-	case map[string]any:
-		value = NewHashValue(v)
-	case map[string]struct{}:
-		value = NewSetValue(v)
-	case map[string]float64:
-		value = NewZSetValue(v)
-	default:
-		// Fallback to string representation
-		value = NewStringValue(fmt.Sprintf("%v", rawValue))
-	}
-	s.data[dbIndex][key] = *value
-	return true, nil
-}
-
-type SetOptions struct {
-	NX bool // Only set if key does not exist
-	XX bool // Only set if key exists
-	EX int  // Expire time in seconds
-	PX int  // Expire time in milliseconds
-}
-
-func parseSetOptions(args []string) (*SetOptions, error) {
-	options := &SetOptions{}
-	i := 0
-	for i < len(args) {
-		switch strings.ToUpper(args[i]) {
-		case "NX":
-			options.NX = true
-			i++
-		case "XX":
-			options.XX = true
-			i++
-		case "EX":
-			if i+1 >= len(args) {
-				return nil, fmt.Errorf("missing value for EX option")
-			}
-			seconds, err := strconv.Atoi(args[i+1])
-			if err != nil {
-				return nil, fmt.Errorf("invalid value for EX option")
-			}
-			options.EX = seconds
-			i += 2
-		case "PX":
-			if i+1 >= len(args) {
-				return nil, fmt.Errorf("missing value for PX option")
-			}
-			milliseconds, err := strconv.Atoi(args[i+1])
-			if err != nil {
-				return nil, fmt.Errorf("invalid value for PX option")
-			}
-			options.PX = milliseconds
-			i += 2
-		default:
-			return nil, fmt.Errorf("unknown option: %s", args[i])
-		}
-	}
-	if options.NX && options.XX {
-		return nil, fmt.Errorf("ERR syntax error")
-	}
-	return options, nil
-}
-
 // Get gets the value for a key
-func (s *Store) Get(dbIndex int, key string) (interface{}, bool) {
+func (s *Store) Get(dbIndex int, key string) (*Value, bool) {
 	s.mu.RLock()
 	defer s.mu.RUnlock()
 	value, ok := s.data[dbIndex][key]
-	if value.IsExpired() {
+	if !ok {
 		return nil, false
 	}
-	if !ok {
+	if value != nil && value.IsExpired() {
 		return nil, false
 	}
 	return value, ok
@@ -316,7 +230,7 @@ func (s *Store) Incr(dbIndex int, key string) (int, error) {
 
 	value, ok := s.data[dbIndex][key]
 	if !ok {
-		value = Value{Data: "0", Type: TypeString}
+		value = &Value{Data: "0", Type: TypeString}
 	}
 	if value.Type != TypeString {
 		return 0, ErrNotInteger
@@ -340,7 +254,7 @@ func (s *Store) Decr(dbIndex int, key string) (int, error) {
 
 	value, ok := s.data[dbIndex][key]
 	if !ok {
-		value = Value{Data: "0", Type: TypeString}
+		value = &Value{Data: "0", Type: TypeString}
 	}
 	if value.Type != TypeString {
 		return 0, ErrNotInteger
@@ -387,7 +301,7 @@ func (s *Store) LPush(dbIndex int, key string, values ...any) int {
 
 	value, ok := s.data[dbIndex][key]
 	if !ok {
-		s.data[dbIndex][key] = *NewListValue(values)
+		s.data[dbIndex][key] = NewListValue(values)
 		return len(values)
 	}
 	list, _ := value.AsList()
@@ -409,7 +323,7 @@ func (s *Store) RPush(dbIndex int, key string, values ...any) int {
 
 	value, ok := s.data[dbIndex][key]
 	if !ok {
-		s.data[dbIndex][key] = *NewListValue(values)
+		s.data[dbIndex][key] = NewListValue(values)
 		return len(values)
 	}
 	list, _ := value.AsList()
