+++ internal/core/store/store_utils.go
@@ -1,16 +1,5 @@
 package store
 
-// // isExpired checks if a key has expired
-// func (s *Store) isExpired(dbIndex int, key string) bool {
-// 	if exp, exists := s.expires[dbIndex][key]; exists {
-// 		if time.Now().After(exp) {
-// 			s.delKey(dbIndex, key)
-// 			return true
-// 		}
-// 	}
-// 	return false
-// }
-
 // delKey deletes a key from the store and its expiration
 func (s *Store) delKey(dbIndex int, key string) {
 	delete(s.data[dbIndex], key)
@@ -18,5 +7,5 @@ func (s *Store) delKey(dbIndex int, key string) {
 
 // flushDb flushes the database
 func (s *Store) flushDb(dbIndex int) {
-	s.data[dbIndex] = make(map[string]Value)
+	s.data[dbIndex] = make(map[string]*Value)
 }
