+++ internal/persistence/rdb/rdb_test.go
@@ -37,16 +37,14 @@ func TestSaveLoadSnapshot(t *testing.T) {
 	}
 
 	// Verify Key1 exists before it expires
-	valInterface, ok := newStore.Get(dbIndex, "Key1")
-	value := valInterface.(store.Value)
+	value, ok := newStore.Get(dbIndex, "Key1")
 	valStr := value.Data.(string)
 	if !ok || valStr != "Value1" {
 		t.Fatalf("Expected Value1, got %s", valStr)
 	}
 
 	// Verify Key2 exists before it expires
-	valInterface, ok = newStore.Get(dbIndex, "Key2")
-	value = valInterface.(store.Value)
+	value, ok = newStore.Get(dbIndex, "Key2")
 	valStr = value.Data.(string)
 	if !ok || valStr != "Value2" {
 		t.Fatalf("Expected Value2, got %s", valStr)
